fix: moved test packages to internal and clear out unused members
