+++ pkg/internal/test/mocks/sessionmanager.go
