+++ pkg/internal/test/mocks/wallet.go
