+++ pkg/internal/test/testutils/request.go
@@ -1,4 +1,4 @@
-package assert
+package testutils
 
 import (
 	"fmt"
@@ -34,40 +34,10 @@ func InternalServerError(t *testing.T, res *http.Response) {
 	require.Equal(t, http.StatusInternalServerError, res.StatusCode)
 }
 
-// MissingRequestIDError checks if the response body contains the "missing request ID" error.
-func MissingRequestIDError(t *testing.T, res *http.Response) {
-	errString := readBody(t, res)
-	require.Equal(t, "missing request ID", errString)
-}
-
-// UnableToVerifySignatureError checks if the response body contains the "unable to verify signature" error.
-func UnableToVerifySignatureError(t *testing.T, res *http.Response) {
-	errString := readBody(t, res)
-	require.Contains(t, errString, "unable to verify signature")
-}
-
-// SessionNotFoundError check if the response body contain the "session not found" error.
-func SessionNotFoundError(t *testing.T, res *http.Response) {
-	errString := readBody(t, res)
-	require.Equal(t, "session not found", errString)
-}
-
-// SessionNotAuthenticatedError check if the response body contain the "session not authenticated" error.
-func SessionNotAuthenticatedError(t *testing.T, res *http.Response) {
-	errString := readBody(t, res)
-	require.Equal(t, "session not authenticated", errString)
-}
-
 // MissingHeaderError check if the response body contain the "missing X header" error.
 func MissingHeaderError(t *testing.T, res *http.Response, header string) {
 	errString := readBody(t, res)
 	require.Contains(t, fmt.Sprintf("missing %s header\n", header), errString)
-}
-
-// InvalidHeaderError check if the response body contain the "invalid X header" error.
-func InvalidHeaderError(t *testing.T, res *http.Response, header string) {
-	errString := readBody(t, res)
-	require.Equal(t, fmt.Sprintf("invalid %s header", header), errString)
 }
 
 // ReadBodyForTest reads and returns the response body for testing purposes
