+++ pkg/internal/test/testutils/response.go
@@ -1,4 +1,4 @@
-package assert
+package testutils
 
 import (
 	"net/http"
